+++ src/app/projects/projects-routing.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { NewprojectComponent } from './newproject/newproject.component';
+import { NewprojectComponent } from './projects/newproject/newproject.component';
 import { ProjectsComponent } from './projects/projects.component';
-import { ViewerprojectComponent } from './viewerproject/viewerproject.component';
+import { ViewerprojectComponent } from './projects/viewerproject/viewerproject.component';
 
 const routes: Routes = [
   {
