+++ src/app/projects/projects.module.ts
@@ -2,10 +2,10 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 
-import { NewprojectComponent } from './newproject/newproject.component';
+import { NewprojectComponent } from './projects/newproject/newproject.component';
 import { ProjectsRoutingModule } from './projects-routing.module';
 import { ProjectsComponent } from './projects/projects.component';
-import { ViewerprojectComponent } from './viewerproject/viewerproject.component';
+import { ViewerprojectComponent } from './projects/viewerproject/viewerproject.component';
 
 
 @NgModule({
