+++ src/app/projects/projects/newproject/newproject.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
-import { environment } from '../../../environments/environment';
-import { Project } from '../models/project.model';
+import { environment } from '../../../../environments/environment';
+import { Project } from '../../models/project.model';
 
 @Component({
   selector: 'app-newproject',
