+++ src/app/projects/projects/projects.component.html
@@ -1,17 +1,21 @@
-<header class="sticky">
-  <a routerLink="/projects/new"
-     class="button"><span>new</span></a>
-</header>
-<div class="container">
-  <ul *ngIf="projects.length>0; else empty">
-    <li *ngFor="let project of projects">
-      <span>{{ project.name }}</span>
-      <a [routerLink]="['project',project.id]"
-         class="button"><span>view</span></a>
-    </li>
-  </ul>
-  <ng-template #empty>
-    <i><strong>No data yet</strong></i>
-  </ng-template>
+<div class="row">
+  <div class="card col-sm-6">
+    <ul *ngIf="projects.length>0; else empty">
+      <li *ngFor="let project of projects">
+        <span>{{ project.name }}</span>
+        <a [routerLink]="['project',project.id]"
+           class="button"><span>view</span></a>
+      </li>
+    </ul>
+    <ng-template #empty>
+      <i><strong>No data yet</strong></i>
+    </ng-template>
+    <div>
+      <a routerLink="/projects/new"
+         class="button primary"><span>new</span></a>
+    </div>
+  </div>
+  <div class="col-sm-6">
+    <router-outlet></router-outlet>
+  </div>
 </div>
-<router-outlet></router-outlet>
