+++ src/app/projects/projects/viewerproject/viewerproject.component.ts
